+++ wp-content/languages/plugins/wowmall-shortcodes-ru_RU.po
