+++ wp-content/plugins/wowmall-shortcodes/assets/css/compare.css
@@ -1,1 +1,108 @@
-.entry-content .wowmall-compare-list.loading{position:relative}.entry-content .wowmall-compare-list.loading:after{position:absolute;width:100%;height:100%;background:rgba(255,255,255,0.5);top:0;left:0;content:''}.entry-content .wowmall-compare-list.loading .wowmall-compare-loader{display:block}.entry-content .wowmall-compare-list .wowmall-compare-loader{position:absolute;border-top:5px solid transparent;border-right:5px solid transparent;border-bottom:5px solid transparent;border-left:5px solid #fc6f38;margin-top:-45px;left:50%;top:50%;margin-left:-45px;transform:translateZ(0);animation:load8 1.1s infinite linear;display:none;z-index:2}.entry-content .wowmall-compare-loader,.entry-content .wowmall-compare-loader:after{border-radius:50%;width:90px;height:90px}@keyframes load8{0%{transform:rotate(0deg)}100%{transform:rotate(360deg)}}.entry-content .wowmall-compare-table{width:100%;min-width:100%;margin:23px 0 25px}.entry-content .wowmall-compare-table.tablesaw{min-width:0}.entry-content .wowmall-compare-table .wowmall-compare-row{background:none}.entry-content .wowmall-compare-table .wowmall-compare-cell{width:auto;position:relative;border-top:1px solid #e4e4e4;border-bottom:1px solid #e4e4e4;padding:30px 0;vertical-align:top;max-width:400px}@media (min-width: 768px){.entry-content .wowmall-compare-table .wowmall-compare-cell{border-left:1px solid #e4e4e4;padding:30px 40px}}.entry-content .wowmall-compare-table .wowmall-compare-cell p{margin:0;padding:0}.entry-content .wowmall-compare-table .wowmall-compare-cell .wowmall-compare-remove{color:#a3a3a3;font-size:1.111111111111111rem;margin:0 0 11px}.entry-content .wowmall-compare-table .wowmall-compare-cell .wowmall-compare-remove span{cursor:pointer;display:inline-block}.entry-content .wowmall-compare-table .wowmall-compare-cell .wowmall-compare-remove span:before{margin:0}.entry-content .wowmall-compare-table .wowmall-compare-cell .product-title{white-space:nowrap;overflow:hidden;text-overflow:ellipsis;margin:4px 0 0}.woocommerce .entry-content .wowmall-compare-table .wowmall-compare-cell .product-status{margin:10px 0 2px}.entry-content .wowmall-compare-table .wowmall-compare-cell .price-wrapper{margin:0 0 17px;min-height:29px}.entry-content .wowmall-compare-table .wowmall-compare-cell .price{font-weight:700;font-size:1.444444444444444rem;display:flex;line-height:1.15}.entry-content .wowmall-compare-table .wowmall-compare-cell .price del{order:1;margin-right:15px}.entry-content .wowmall-compare-table .wowmall-compare-cell .price ins{padding:0;background:none;margin-right:15px;color:#de2e2e}.entry-content .wowmall-compare-table .wowmall-compare-heading-cell{vertical-align:top;padding:32px 0;width:200px;border-top:1px solid #e4e4e4;border-bottom:1px solid #e4e4e4;font-weight:700;font-size:.8888888888888889rem;text-transform:uppercase;letter-spacing:.06em}@media (max-width: 767px){.entry-content .wowmall-compare-table .wowmall-compare-heading-cell{display:none}}
+.entry-content .wowmall-compare-list.loading {
+  position: relative; }
+  .entry-content .wowmall-compare-list.loading:after {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background: rgba(255, 255, 255, 0.5);
+    top: 0;
+    left: 0;
+    content: ''; }
+  .entry-content .wowmall-compare-list.loading .wowmall-compare-loader {
+    display: block; }
+.entry-content .wowmall-compare-list .wowmall-compare-loader {
+  position: absolute;
+  border-top: 5px solid transparent;
+  border-right: 5px solid transparent;
+  border-bottom: 5px solid transparent;
+  border-left: 5px solid #fc6f38;
+  margin-top: -45px;
+  left: 50%;
+  top: 50%;
+  margin-left: -45px;
+  transform: translateZ(0);
+  animation: load8 1.1s infinite linear;
+  display: none;
+  z-index: 2; }
+.entry-content .wowmall-compare-loader,
+.entry-content .wowmall-compare-loader:after {
+  border-radius: 50%;
+  width: 90px;
+  height: 90px; }
+@keyframes load8 {
+  0% {
+    transform: rotate(0deg); }
+  100% {
+    transform: rotate(360deg); } }
+.entry-content .wowmall-compare-table {
+  width: 100%;
+  min-width: 100%;
+  margin: 23px 0 25px; }
+  .entry-content .wowmall-compare-table.tablesaw {
+    min-width: 0; }
+  .entry-content .wowmall-compare-table .wowmall-compare-row {
+    background: none; }
+  .entry-content .wowmall-compare-table .wowmall-compare-cell {
+    width: auto;
+    position: relative;
+    border-top: 1px solid #e4e4e4;
+    border-bottom: 1px solid #e4e4e4;
+    padding: 30px 0;
+    vertical-align: top;
+    max-width: 400px; }
+    @media (min-width: 768px) {
+      .entry-content .wowmall-compare-table .wowmall-compare-cell {
+        border-left: 1px solid #e4e4e4;
+        padding: 30px 40px; } }
+    .entry-content .wowmall-compare-table .wowmall-compare-cell p {
+      margin: 0;
+      padding: 0; }
+    .entry-content .wowmall-compare-table .wowmall-compare-cell .wowmall-compare-remove {
+      color: #a3a3a3;
+      font-size: 1.111111111111111rem;
+      margin: 0 0 11px; }
+      .entry-content .wowmall-compare-table .wowmall-compare-cell .wowmall-compare-remove span {
+        cursor: pointer;
+        display: inline-block; }
+        .entry-content .wowmall-compare-table .wowmall-compare-cell .wowmall-compare-remove span:before {
+          margin: 0; }
+    .entry-content .wowmall-compare-table .wowmall-compare-cell .product-title {
+      margin: 4px 0 0; }
+      @media (min-width: 768px) {
+        .entry-content .wowmall-compare-table .wowmall-compare-cell .product-title {
+          white-space: nowrap;
+          overflow: hidden;
+          text-overflow: ellipsis; } }
+    .woocommerce .entry-content .wowmall-compare-table .wowmall-compare-cell .product-status {
+      margin: 10px 0 2px; }
+    .entry-content .wowmall-compare-table .wowmall-compare-cell .price-wrapper {
+      margin: 0 0 17px;
+      min-height: 29px; }
+    .entry-content .wowmall-compare-table .wowmall-compare-cell .price {
+      font-weight: 700;
+      font-size: 1.444444444444444rem;
+      display: flex;
+      line-height: 1.15; }
+      .entry-content .wowmall-compare-table .wowmall-compare-cell .price del {
+        order: 1;
+        margin-right: 15px; }
+      .entry-content .wowmall-compare-table .wowmall-compare-cell .price ins {
+        padding: 0;
+        background: none;
+        margin-right: 15px;
+        color: #de2e2e; }
+  .entry-content .wowmall-compare-table .wowmall-compare-heading-cell {
+    vertical-align: top;
+    padding: 32px 0;
+    width: 200px;
+    border-top: 1px solid #e4e4e4;
+    border-bottom: 1px solid #e4e4e4;
+    font-weight: 700;
+    font-size: .8888888888888889rem;
+    text-transform: uppercase;
+    letter-spacing: .06em; }
+    @media (max-width: 767px) {
+      .entry-content .wowmall-compare-table .wowmall-compare-heading-cell {
+        display: none; } }
+
+/*# sourceMappingURL=compare.css.map */
