+++ wp-content/plugins/wowmall-shortcodes/assets/css/wishlist.scss
