+++ wp-content/plugins/wowmall-shortcodes/assets/js/compare.js
@@ -17,7 +17,7 @@
 						wowmallParams.ajax_url,
 						{
 							action: 'wowmall_compare_button',
-							url   : encodeURIComponent(button.closest( '.product' ).find( '.woocommerce-LoopProduct-link' ).attr('href'))
+							url   : $(document.body).hasClass( 'single-product' ) ? encodeURIComponent(location.href) : encodeURIComponent(button.closest( '.product' ).find( '.woocommerce-LoopProduct-link' ).attr('href'))
 						},
 						function ( response ) {
 							button.removeClass( wowmallCompareLoadingClass );
