+++ wp-content/plugins/wowmall-shortcodes/inc/widgets/class-wc-widget-layered-nav-filters.php
@@ -39,7 +39,7 @@ class Wowmall_Shortcodes_WC_Widget_Layered_Nav_Filters extends WC_Widget_Layered
 		$min_price          = isset( $_GET['min_price'] ) ? wc_clean( $_GET['min_price'] ) : 0;
 		$max_price          = isset( $_GET['max_price'] ) ? wc_clean( $_GET['max_price'] ) : 0;
 		$rating_filter      = isset( $_GET['rating_filter'] ) ? array_filter( array_map( 'absint', explode( ',', $_GET['rating_filter'] ) ) ) : array();
-		$base_link          = $this->get_page_base_url();
+		$base_link          = $this->get_current_page_url();
 
 		if ( 0 < count( $_chosen_attributes ) || 0 < $min_price || 0 < $max_price || ! empty( $rating_filter ) ) {
