+++ wp-content/themes/wowmall-child/languages/ru_RU.po
