+++ wp-content/themes/wowmall-child/style.css
@@ -1,0 +1,13 @@
+/*
+Theme Name: Slamdunk Shop
+Theme URI: https://slamdunk.shop
+Template: wowmall
+Author: Slamdunk Shop
+Author URI: https://slamdunk.shop
+Description: Дочерняя тема Wowmall
+Tags: one-column,two-columns,left-sidebar,accessibility-ready,custom-background,custom-colors,custom-header,custom-menu,editor-style,featured-images,flexible-header,microformats,post-formats,rtl-language-support,sticky-post,threaded-comments,translation-ready,theme-options
+Version: 1.0.1528020389
+Updated: 2018-06-03 18:06:29
+
+*/
+
