+++ wp-content/themes/wowmall-child/style.ctcbackup.css
@@ -1,0 +1,13 @@
+/*
+Theme Name: Wowmall Child
+Theme URI: http://wowmall.tonytemplates.com/
+Template: wowmall
+Author: Etheme
+Author URI: https://themeforest.net/user/etheme/portfolio
+Description: Powerfull WordPress WooCommerce theme
+Tags: one-column,two-columns,left-sidebar,accessibility-ready,custom-background,custom-colors,custom-header,custom-menu,editor-style,featured-images,flexible-header,microformats,post-formats,rtl-language-support,sticky-post,threaded-comments,translation-ready,theme-options
+Version: 1.5.0.1528017932
+Updated: 2018-06-03 17:25:32
+
+*/
+
